+++ display_metrics_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+* Multi-display, foldable devices support
+* Implemented *getDisplays* method
+* Updated *display_metrics_platform_interface* dependency (1.1.0)
+
 ## 1.0.3
 * Updated documentation